Externalize connect-protocol
